+++ mxsniff/__init__.py
@@ -89,6 +89,7 @@ provider_domains = WildcardDomainDict()
 for name, data in providers.items():
     for domain in data['mx']:
         provider_domains[domain] = name
+del name, data, domain
 
 
 class MXLookupException(Exception):
@@ -117,7 +118,7 @@ def get_domain(email_or_domain):
     return domain
 
 
-def mxsniff(email_or_domain, verbose=False, ignore_errors=False):
+def mxsniff(email_or_domain, ignore_errors=False, cache=None):
     """
     Lookup MX records for a given email address, URL or domain name and identify the email service provider(s)
     from an internal list of known service providers.
@@ -125,26 +126,30 @@ def mxsniff(email_or_domain, verbose=False, ignore_errors=False):
     :param str email_or_domain: Email, domain or URL to lookup
     :return: Identified service provider, or a list if there's more than one (in unusual circumstances)
 
-    >>> mxsniff('example.com')
-    'nomx'
-    >>> mxsniff('__invalid_domain_name__.com')
-    'nomx'
-    >>> mxsniff('example@gmail.com')
-    'google-gmail'
-    >>> mxsniff('https://google.com/')
-    'google-apps'
-    >>> sorted(mxsniff('google.com', verbose=True).items())
-    [('match', ['google-apps']), ('mx', [(10, 'aspmx.l.google.com'), (20, 'alt1.aspmx.l.google.com'), (30, 'alt2.aspmx.l.google.com'), (40, 'alt3.aspmx.l.google.com'), (50, 'alt4.aspmx.l.google.com')]), ('name', 'google.com')]
+    >>> mxsniff('example.com')['match']
+    ['nomx']
+    >>> mxsniff('__invalid_domain_name__.com')['match']
+    ['nomx']
+    >>> mxsniff('example@gmail.com')['match']
+    ['google-gmail']
+    >>> sorted(mxsniff('https://google.com/').items())
+    [('domain', 'google.com'), ('match', ['google-apps']), ('mx', [(10, 'aspmx.l.google.com'), (20, 'alt1.aspmx.l.google.com'), (30, 'alt2.aspmx.l.google.com'), (40, 'alt3.aspmx.l.google.com'), (50, 'alt4.aspmx.l.google.com')]), ('mx_tld', ['google.com']), ('query', 'https://google.com/')]
     """
     domain = get_domain(email_or_domain)
+    if cache and domain in cache:
+        return cache[domain]
 
     result = []
+    tld = []
 
     try:
         answers = []  # Default value in case of verbose mode where an error occurs
         answers = sorted([(rdata.preference, rdata.exchange.to_text(omit_final_dot=True).lower())
             for rdata in dns.resolver.query(domain, 'MX')])
         for preference, exchange in answers:
+            rdomain = tldextract.extract(exchange).registered_domain
+            if rdomain not in tld:
+                tld.append(rdomain)
             provider = provider_domains.get(exchange)
             if provider and provider not in result:
                 result.append(provider)
@@ -158,39 +163,43 @@ def mxsniff(email_or_domain, verbose=False, ignore_errors=False):
                 exc=e.__class__.__name__, error=text_type(e), domain=domain))
 
     if not result:
-        # Check for self-hosted email servers, identified with the label 'self'
-        rdomain = tldextract.extract(domain).registered_domain
-        for preference, exchange in answers:
-            if tldextract.extract(exchange).registered_domain == rdomain:
-                result.append('self')
-                break
+        # Check for self-hosted email servers; identify them with the label 'self'
+        if tldextract.extract(domain).registered_domain in tld:
+            result.append('self')
         if not result:
             if answers:
                 result.append('unknown')  # We don't know this one's provider
             else:
                 result.append('nomx')  # This domain has no mail servers
 
-    if verbose:
-        return {'name': email_or_domain, 'match': result, 'mx': answers}
-    else:
-        if len(result) == 1:
-            return result[0]
-        else:
-            return result
+    result = {'query': email_or_domain, 'domain': domain, 'match': result, 'mx': answers, 'mx_tld': tld}
+    if cache:
+        cache[domain] = result
+    return result
 
 
-def mxbulksniff(items, verbose=False, ignore_errors=True):
+def mxbulksniff(items, ignore_errors=True):
     """
     Identify the email service provider of a large set of domains or emails, caching to avoid
     repeat queries. Returns a generator that yields one item at a time
 
-    >>> list(mxbulksniff(['example.com', 'google.com', 'http://www.google.com']))
-    [('example.com', 'nomx'), ('google.com', 'google-apps'), ('http://www.google.com', 'google-apps')]
+    >>> [(i['query'], i['match']) for i in mxbulksniff(['example.com', 'google.com', 'http://www.google.com'])]
+    [('example.com', ['nomx']), ('google.com', ['google-apps']), ('http://www.google.com', ['google-apps'])]
     """
-    domain_cache = {}
+    cache = {}
     for i in items:
-        domain = get_domain(i)
-        yield i, domain_cache[domain] if domain in domain_cache else mxsniff(domain, verbose, ignore_errors)
+        yield mxsniff(i, ignore_errors, cache)
+
+
+def multiprocess_mxsniff(email_or_domain, ignore_errors=True):
+    """
+    mxsniff wrapper that converts a KeyboardInterrupt into an Exception. Required because
+    of Python bug #8296: http://bugs.python.org/issue8296
+    """
+    try:
+        return mxsniff(email_or_domain, ignore_errors)
+    except KeyboardInterrupt:
+        raise Exception("KeyboardInterrupt")
 
 
 def main_internal(args, name='mxsniff'):
@@ -201,6 +210,9 @@ def main_internal(args, name='mxsniff'):
     example@gmail.com: google-gmail
     """
     import argparse
+    import json
+    from multiprocessing import Pool
+
     parser = argparse.ArgumentParser(
         prog=name,
         description='Identify email service providers given an email address, URL or domain name',
@@ -212,8 +224,17 @@ def main_internal(args, name='mxsniff'):
     parser.add_argument('-i', '--ignore-errors', action='store_true',
         help="ignore DNS lookup errors and continue with next item")
     args = parser.parse_args(args)
-    for item, provider in mxbulksniff(args.names, verbose=args.verbose, ignore_errors=args.ignore_errors):
-        print("{item}: {provider}".format(item=item, provider=provider))
+
+    pool = Pool(processes=10)
+    it = pool.imap(multiprocess_mxsniff, args.names, 10)
+    try:
+        for result in it:
+            if args.verbose:
+                print(json.dumps(result)) + ','
+            else:
+                print("{item}: {provider}".format(item=result['query'], provider=', '.join(result['match'])))
+    except KeyboardInterrupt:
+        pool.terminate()
 
 
 def main():