+++ CHANGES.rst
@@ -4,7 +4,9 @@
 * Added support for wildcards in domain names
 * Additional providers
 * Detect self-hosted email servers
-* Track MX TLDs in verbose mode
+* Remove verbose mode in the mxsniff function; always verbose now
+* Track MX TLDs
+* In the command line version, run queries in a multiprocess pool
 
 0.2.1
 =====